refactor: address code-review findings
- useUserStats: return only the fields consumers use instead of spreading
  ...rest from useUsers, so React Query v5's tracked-render optimization
  isn't defeated (the card no longer re-renders on every background tick).
- formatDate: hoist the helper into @/lib/formatters (alongside
  currentMonthRange/formatCurrency) and consume it from dashboard and
  reports, removing the two byte-identical per-feature copies.
