+++ src/features/dashboard/financial-overview.tsx
@@ -8,10 +8,10 @@ import { Alert, AlertDescription } from '@/components/ui/alert'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import type { ChartConfig } from '@/components/ui/chart'
 import { useDonationReport, useExpenseReport } from '@/features/reports'
-import { currentMonthRange } from '@/lib/formatters'
+import { currentMonthRange, formatDate } from '@/lib/formatters'
 import { getProblemMessage } from '@/lib/get-problem-message'
 import { ComparisonBarChart } from './comparison-bar-chart'
-import { formatDate, previousRange } from './financial-overview.utils'
+import { previousRange } from './financial-overview.utils'
 import { StatCard } from './stat-card'
 import { useBalance } from './use-dashboard-data'
