+++ src/features/dashboard/use-user-stats.ts
@@ -1,9 +1,11 @@
 import { useUsers } from '@/features/users'
 
 export function useUserStats() {
-  const { data, ...rest } = useUsers({ page: 0, size: 1 })
+  const { data, isLoading, isSuccess, error } = useUsers({ page: 0, size: 1 })
   return {
-    ...rest,
     data: data ? { totalUsers: data.page?.totalElements } : undefined,
+    isLoading,
+    isSuccess,
+    error,
   }
 }
