--- src/features/reports/format-date.ts
@@ -1,5 +1,0 @@
-import dayjs from 'dayjs'
-
-export function formatDate(d: Date) {
-  return dayjs(d).format('YYYY-MM-DD')
-}
