+++ src/lib/formatters.ts
@@ -7,6 +7,10 @@ export function formatCurrency(amount: number): string {
   }).format(amount)
 }
 
+export function formatDate(d: Date): string {
+  return dayjs(d).format('YYYY-MM-DD')
+}
+
 export function currentMonthRange(): { from: Date; to: Date } {
   return {
     from: dayjs().startOf('month').toDate(),
